rename event_handler.js to workerMessenger.js
